prepare_basic_rootfs: fix download, upgrade, repack rootfs
the exit status after such a check [[ ${EVALPIPE[0]} -ne 0 ]] &&
design will have the opposite value. And the exit status of
the function will always be "1" which is an error in itself.
Although there are no errors inside the function.

Now to download rootfs from the Armbian site, you must set the key
DOWNLOAD_ROOTFS=yes and the archive file must be reachable for download.

The archive compression ratio is set to 9 by default, unless another
value of the K_ZST key is set. Great thanks @150balbes